src/core/shared/domain/events/domain-event.mediator.ts
import EventEmitter2 from 'eventemitter2';
import { AggregateRoot } from '@core/shared/domain/entity/aggregate-root';
import { IDomainEventHandler } from '@core/shared/application/domain-event-handler.interface';

export class DomainEventMediator {
  constructor(private eventEmitter: EventEmitter2) {}

  register(event: string, handler: IDomainEventHandler) {
    this.eventEmitter.on(event, handler.handle);
  }

  async publish(aggregateRoot: AggregateRoot) {
    for (const event of aggregateRoot.getUncommittedEvents()) {
      const eventClassName = event.constructor.name;
      aggregateRoot.markEventAsDispatched(event);
      await this.eventEmitter.emitAsync(eventClassName, event);
    }
  }

  async publishIntegrationEvents(aggregateRoot: AggregateRoot) {
    for (const event of aggregateRoot.events) {
      const integrationEvent = event.getIntegrationEvent?.();
      if (!integrationEvent) continue;
      await this.eventEmitter.emitAsync(
        integrationEvent.constructor.name,
        integrationEvent,
      );
    }
  }
}
